mediapipe/model_maker/requirements.txt
absl-py
mediapipe>=0.10.0
numpy
opencv-python
tensorflow>=2.10
tensorflow-addons
tensorflow-datasets
tensorflow-hub
tf-models-official>=2.13.1
